refactor(getCarNames): 자동차 이름 입력 로직을 모듈로 분리
자동차 이름을 입력받는 메소드 getCarNames 를 App이 아닌 다른 클래스의 메소드로 분리 : Game/UserInputCarNames.js

테스트 실황
Test Suites: 2 passed, 2 total
Tests:       7 passed, 7 total
Snapshots:   0 total
Time:        0.684 s, estimated 1 s
